Rename pages and update navigation
