+++ index.html
@@ -15,12 +15,12 @@
       <nav id="sidebar">
         <ul>
           <li><a href="/" data-navigo>Home</a></li>
-          <li><a href="/counter" data-navigo>Counter</a></li>
-          <li><a href="/three_cube" data-navigo>Cube Scene</a></li>
-          <li><a href="/three_sphere" data-navigo>Sphere Scene</a></li>
-          <li><a href="/function_plot" data-navigo>3D Function Plot</a></li>
-          <li><a href="/formula_graph" data-navigo>Formula Graph</a></li>
-          <li><a href="/riemann" data-navigo>Riemann Zeros</a></li>
+          <li><a href="/counter" data-navigo>Simple Counter</a></li>
+          <li><a href="/three_cube" data-navigo>Cube Demo</a></li>
+          <li><a href="/three_sphere" data-navigo>Sphere Demo</a></li>
+          <li><a href="/function_plot" data-navigo>Complex Plot Demo</a></li>
+          <li><a href="/formula_graph" data-navigo>Formula Grapher</a></li>
+          <li><a href="/riemann" data-navigo>Riemann Zeta Zeros</a></li>
         </ul>
       </nav>
       <div id="content">Loading...</div>
